update와 delete에 대한 권한 설정

+++ posts/views.py
@@ -34,12 +34,16 @@ def list(request):
 
 def delete(request, post_id):
     post = get_object_or_404(Post, id=post_id)
+    if post.user != request.user:
+        return render('posts:list')
     post.delete()
     return redirect('posts:list')
 
 
 def update(request, post_id):
     post = get_object_or_404(Post, id=post_id)
+    if post.user != request.user:
+        return redirect('posts:list')
     if request.method == 'POST':
         form = PostModelForm(request.POST, instance=post)
         if form.is_valid():
